.ipynb_checkpoints/requirements-checkpoint.txt
numpy
Pillow
scikit-learn
matplotlib
requests
beautifulsoup4
cloudscraper